+++ Assets/CHANGELOG.md
@@ -1,2 +1,11 @@
+# v1.1
+- Added the TagQuery, a static script that can handle all the tags you added in the scene.
+```
+TagQuery.FindObject("tagName");
+
+It uses a dictionary to loop through the objects faster than the findobjectsoftype.
+```
+- Folder organization
+
 # V1.00
 - Release with all the base functions working